git.sh: NOUPLOAD feature

+++ git.sh
@@ -86,7 +86,7 @@ for dist in $DISTROS ; do
 done
 
 popd
-if [ -n "$PPA" ]; then
+if [ -n "$PPA" ] && [ -z "$NOUPLOAD" ]; then
 	dput ppa:ernstp/"$PPA" ${PACKAGE_NAME}_*_source.changes
 	rm -vf *.dsc *.build *.buildinfo *.changes *.upload *.tar.gz *.tar.xz
 fi
